Move check_components types
