+++ crates/macros/cgp-macro-core/src/types/check_components/check_entry.rs
@@ -1,0 +1,11 @@
+use proc_macro2::Span;
+use syn::Type;
+
+use crate::types::generics::ImplGenerics;
+
+pub struct CheckEntry {
+    pub component_type: Type,
+    pub component_params: Option<Type>,
+    pub span: Span,
+    pub generics: ImplGenerics,
+}
