+++ crates/macros/cgp-macro-core/src/types/check_components/type_with_generics.rs
@@ -1,0 +1,24 @@
+use syn::Type;
+use syn::parse::{Parse, ParseStream};
+use syn::token::Lt;
+
+use crate::types::generics::ImplGenerics;
+
+pub struct TypeWithGenerics {
+    pub ty: Type,
+    pub generics: ImplGenerics,
+}
+
+impl Parse for TypeWithGenerics {
+    fn parse(input: ParseStream) -> syn::Result<Self> {
+        let generics = if input.peek(Lt) {
+            input.parse()?
+        } else {
+            ImplGenerics::default()
+        };
+
+        let ty = input.parse()?;
+
+        Ok(Self { ty, generics })
+    }
+}
