+++ crates/macros/cgp-macro-lib/src/entrypoints/check_components.rs
@@ -1,12 +1,12 @@
+use cgp_macro_core::types::check_components::CheckComponentsTables;
 use proc_macro2::TokenStream;
 use quote::{ToTokens, TokenStreamExt};
 use syn::parse2;
 
 use crate::check_components::derive_check_components;
-use crate::parse::CheckComponentsSpecs;
 
 pub fn check_components(body: TokenStream) -> syn::Result<TokenStream> {
-    let spec: CheckComponentsSpecs = parse2(body)?;
+    let spec: CheckComponentsTables = parse2(body)?;
 
     let mut out = TokenStream::new();
