+++ crates/macros/cgp-macro-lib/src/parse/mod.rs
@@ -1,11 +1,9 @@
-mod check_components;
 mod define_preset;
 mod delegate_and_check_components;
 mod delegate_components;
 mod path;
 mod type_spec;
 
-pub use check_components::*;
 pub use define_preset::*;
 pub use delegate_and_check_components::*;
 pub use delegate_components::*;
